+++ RPGEconomy.Application/Abstractions/Services/IMarketService.cs
@@ -7,6 +7,6 @@ public interface IMarketService
 {
     Task<Result<IReadOnlyList<MarketPriceDto>>> GetPricesAsync(int settlementId);
     Task<Result<MarketPriceDto>> GetProductAsync(int settlementId, int productTypeId);
-    Task<Result<MarketPriceDto>> UpdateProductStateAsync(int settlementId, int productTypeId, int supply, int demand);
+    Task<Result<MarketPriceDto>> UpdateProductStateAsync(int settlementId, int productTypeId, decimal supply, decimal demand);
     Task<Result> RegisterProductAsync(int settlementId, int productTypeId, decimal initialPrice);
 }
