+++ RPGEconomy.Domain/Markets/MarketPricePolicy.cs
@@ -6,18 +6,18 @@ internal static class MarketPricePolicy
     private const decimal MinPrice = 0.01m;
     private const decimal MaxTickChangeRatio = 0.5m;
 
-    public static decimal Recalculate(decimal currentPrice, int supply, int demand)
+    public static decimal Recalculate(decimal currentPrice, decimal supply, decimal demand)
     {
         if (currentPrice < MinPrice)
             currentPrice = MinPrice;
 
-        if (supply == 0 && demand == 0)
+        if (supply == 0m && demand == 0m)
             return currentPrice;
 
-        if (supply == 0 && demand > 0)
+        if (supply == 0m && demand > 0m)
             return ApplyFactor(currentPrice, 1m + Sensitivity);
 
-        var ratio = (decimal)demand / supply;
+        var ratio = demand / supply;
         var rawFactor = 1m + Sensitivity * (ratio - 1m);
         var minFactor = 1m - MaxTickChangeRatio;
         var maxFactor = 1m + MaxTickChangeRatio;
